+++ autosubmit/config/basicconfig.py
@@ -20,6 +20,7 @@ import os
 from configparser import ConfigParser
 from pathlib import Path
 from typing import Union
+from autosubmit.log.log import Log
 
 
 class BasicConfig:
@@ -97,7 +98,7 @@ class BasicConfig:
     @staticmethod
     def __read_file_config(file_path):
         """
-        Reads configuration file. If configuration file dos not exist in given path,
+        Reads configuration file. If configuration file does not exist in given path,
         no error is raised. Configuration options also are not required to exist
 
         :param file_path: configuration file to read
@@ -197,21 +198,35 @@ class BasicConfig:
         Reads configuration from .autosubmitrc files, first from /etc., then for user
         directory and last for current path.
         """
-        filename = 'autosubmitrc'
-        if 'AUTOSUBMIT_CONFIGURATION' in os.environ and os.path.exists(os.environ['AUTOSUBMIT_CONFIGURATION']):
-            config_file_path = os.environ['AUTOSUBMIT_CONFIGURATION']
+        filename = "autosubmitrc"
+        dot_filename = f".{filename}"
+        # Check if the environment variable is defined
+        if (
+            "AUTOSUBMIT_CONFIGURATION" in os.environ
+            and Path(os.environ["AUTOSUBMIT_CONFIGURATION"]).exists()
+        ):
+            config_file_path = os.environ["AUTOSUBMIT_CONFIGURATION"]
             # Call read_file_config with the value of the environment variable
             BasicConfig.__read_file_config(config_file_path)
         else:
-            if os.path.exists(os.path.join('', '.' + filename)):
-                BasicConfig.__read_file_config(os.path.join('', '.' + filename))
-            elif os.path.exists(os.path.join(os.path.expanduser('~'), '.' + filename)):
-                BasicConfig.__read_file_config(os.path.join(
-                    os.path.expanduser('~'), '.' + filename))
-            else:
-                BasicConfig.__read_file_config(os.path.join('/etc', filename))
+            user_config_path = Path(Path.cwd(), dot_filename)
+            home_user_config_path = Path(Path.home(), dot_filename)
+            etc_rc_path = Path("/etc", filename)
+            legacy_etc_rc_path = Path("/etc", dot_filename)
 
-            # Check if the environment variable is defined
+            if user_config_path.exists():
+                BasicConfig.__read_file_config(user_config_path)
+            elif home_user_config_path.exists():
+                BasicConfig.__read_file_config(home_user_config_path)
+            else:
+                if legacy_etc_rc_path.exists():
+                    Log.warning(
+                        "The legacy configuration file /etc/.autosubmitrc is deprecated and will be removed in future versions. Please, rename it to /etc/autosubmitrc"
+                    )
+                    BasicConfig.__read_file_config(legacy_etc_rc_path)
+                # Overwrite legacy config
+                if etc_rc_path.exists():
+                    BasicConfig.__read_file_config(etc_rc_path)
 
         BasicConfig._update_config()
         return